+++ packages/discovery/src/discovery/analysis/TemplateService.ts
@@ -1,8 +1,14 @@
 import { existsSync, readFileSync, readdirSync } from 'fs'
 import path, { join } from 'path'
 
+import type { DiscoveryOutput } from '@l2beat/discovery-types'
 import { hashJson } from '@l2beat/shared'
-import { EthereumAddress, type Hash256, type json } from '@l2beat/shared-pure'
+import {
+  assert,
+  EthereumAddress,
+  Hash256,
+  type json,
+} from '@l2beat/shared-pure'
 import {
   flattenFirstSource,
   flatteningHash,
@@ -10,6 +16,7 @@ import {
   sha2_256bit,
 } from '../../flatten/utils'
 import { fileExistsCaseSensitive } from '../../utils/fsLayer'
+import type { DiscoveryConfig } from '../config/DiscoveryConfig'
 import { DiscoveryContract } from '../config/RawDiscoveryConfig'
 import type { ContractSources } from '../source/SourceCodeService'
 import { readJsonc } from '../utils/readJsonc'
@@ -168,6 +175,68 @@ export class TemplateService {
     }
     this.allTemplateHashes = result
     return result
+  }
+
+  // returns reason or undefined
+  discoveryNeedsRefresh(discovery: DiscoveryOutput, config: DiscoveryConfig) {
+    const allTemplateHashes = this.getAllTemplateHashes()
+    const allShapes = this.getAllShapes()
+
+    for (const contract of discovery.contracts) {
+      if (contract.sourceHashes === undefined) {
+        continue
+      }
+      const hashes =
+        contract.sourceHashes.length === 1
+          ? contract.sourceHashes
+          : contract.sourceHashes.slice(1)
+
+      if (hashes.length > 1) {
+        // NOTE(radomski): Diamonds don't really work well with templates right now
+        continue
+      }
+
+      const hash = hashes[0]
+      assert(hash !== undefined)
+      const sourcesHash = Hash256(hash)
+      const matchingTemplates = this.findMatchingTemplatesByHash(
+        sourcesHash,
+        contract.address,
+      )
+
+      if (
+        contract.template !== undefined &&
+        (allShapes[contract.template]?.hashes.length ?? 0) > 0
+      ) {
+        if (config.for(contract.address).extends === undefined) {
+          if (matchingTemplates.length === 0) {
+            return `A contract "${contract.name}" with template "${contract.template}", no longer matches any template`
+          }
+          if (contract.template !== matchingTemplates[0]) {
+            return `A contract "${contract.name}" matches a different template: "${contract.template} -> ${matchingTemplates.join(', ')}"`
+          }
+        }
+      } else if (matchingTemplates.length > 0) {
+        return `A contract "${contract.name}" without template now matches: "${matchingTemplates.join(', ')}"`
+      }
+    }
+
+    if (discovery.configHash !== config.hash) {
+      return 'project config or used template has changed'
+    }
+
+    const outdatedTemplates = []
+    for (const [templateId, templateHash] of Object.entries(
+      discovery.usedTemplates,
+    )) {
+      if (templateHash !== allTemplateHashes[templateId]) {
+        outdatedTemplates.push(templateId)
+      }
+    }
+
+    if (outdatedTemplates.length > 0) {
+      return `template configs has changed: ${outdatedTemplates.join(', ')}`
+    }
   }
 }
