+++ packages/discovery/src/discovery/config/ConfigReader.ts
@@ -3,7 +3,8 @@ import { readFileSync } from 'fs'
 import path from 'path'
 import type { DiscoveryOutput } from '@l2beat/discovery-types'
 
-import { assert, stripAnsiEscapeCodes } from '@l2beat/shared-pure'
+import { createHash } from 'crypto'
+import { assert, Hash160, stripAnsiEscapeCodes } from '@l2beat/shared-pure'
 import chalk from 'chalk'
 import type { ZodError } from 'zod'
 import { fileExistsCaseSensitive } from '../../utils/fsLayer'
@@ -81,6 +82,13 @@ export class ConfigReader {
     const meta = JSON.parse(contents) as unknown as DiscoveryOutput
     assert(meta.chain === chain, 'Chain mismatch in discovered.json')
     return meta
+  }
+
+  readDiscoveryHash(projectName: string, chain: string): Hash160 {
+    const curDiscovery = this.readDiscovery(projectName, chain)
+    const hasher = createHash('sha1')
+    hasher.update(JSON.stringify(curDiscovery))
+    return Hash160(`0x${hasher.digest('hex')}`)
   }
 
   readAllChains(): string[] {
